requirements.txt
adbutils==0.7.1
apkutils2==1.0.0
Appium-Python-Client==0.50
atomicwrites==1.3.0
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
cigam==0.0.3
colorama==0.4.3
decorator==4.4.2
Deprecated==1.2.7
deprecation==2.0.7
enum34==1.1.10
facebook-wda==0.7.2
humanize==2.0.0
idna==2.9
importlib-metadata==1.5.0
logzero==1.5.0
lxml==4.5.0
more-itertools==8.2.0
namedlist==1.7
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
progress==1.5
py==1.8.1
pyelftools==0.26
pyparsing==2.4.6
pytest==5.4.1
pytest-allure-adaptor==1.6.0
pytest-html==2.1.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==8.0
PyYAML==5.3
requests==2.23.0
retry==0.9.2
selenium==3.141.0
six==1.14.0
tornado==6.0.4
uiautomator2==2.5.7
urllib3==1.25.8
wcwidth==0.1.8
weditor==0.4.3
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
